js/runner.go: Run script function and it's eventLoop in parallel
This patch modifies the `runFn()` to run the user defined function
in js and it's eventLoop queue in parallel. This is done by calling
`eventLoop.WaitOnRegistered()` every second.

Previously, the user function was passed as first item to execute in the
eventLoop, and then `WaitOnRegistered()` was being called only after
main function has exited. This means events defined in js will only be
delivered after the main function in the VU has exited.

Using long running VUs to load test `watch()` like gRPC streaming APIs
is not possible with the current behaviour of eventLoop processing after
the VU returns.

Signed-off-by: Prem Kumar <prem.saraswat@grafana.com>